requirements.txt
python-telegram-bot>=20.0
requests==2.31.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
asyncio==3.4.3
SQLAlchemy>=1.4.0
pytz>=2021.1
